+++ CHANGELOG.rst
@@ -6,6 +6,7 @@ dev
 - Added new example: the stochastic Lotka-Volterra model
 - Fix methods.bo.utils.minimize to be strictly within bounds
 - Implemented the Two Stage Procedure, a method of summary-statistics diagnostics  
+- Added the MaxVar acquisition method
 
 0.6.3 (2017-09-28)
 ------------------